Change Adapter Method Create

+++ PixCharge.Integration.Adapter/IPix.cs
@@ -3,6 +3,6 @@
 namespace PixCharge.Integration.Adapter;
 public interface IPix
 {
-    public Charge CreateCharge(long Value, string CorrelationID);
+    public Charge CreateCharge(decimal Value, string CorrelationID);
     public Boolean IsChargeApporve(Guid correlationID);
 }
